py/k8s/python-authentication-webhook-app/requirements.txt
flask
jsonpatch
pyyaml
pymysql
wtforms
requests
python-dotenv
